+++ src/Random/Distribution/UniformRealDistribution.php
@@ -54,6 +54,6 @@ class UniformRealDistribution extends AbstractDistribution
      */
     public function generate(AbstractEngine $engine)
     {
-        return $this->min + ($this->max - $this->min) * $engine->nextFloat();
+        return $this->min + ($this->max - $this->min) * $engine->nextDouble();
     }
 }
